Changes in H4_PadController_Rev_2 and H4_MainController_Rev_1
In H4_PadController_Rev_2:

- Created FIXED_I2C_DIR to stablish a fixed I2c direction during programming to avoid EEPROM corruption

In H4_MainController_Rev_1:

- Created voiceSituationReport();
- Fixed i2c Directions
- Fixed I2c Voice and LED Commands to happen first after push

+++ 02_Elements/H4_PadController_Rev_2/03_SW/H4_PadController_Rev_2/H4_Configuration.h
@@ -45,21 +45,21 @@ class H4_Configuration
     void set_defaultI2cDirConfiguration();
     // Structure to easily manage all the parameters
     struct storedConfiguration{
-        int i2cDIR_Stored;
-        int primaryColor;
-        int secondaryColor;
-        int disconnectedColor;
-        int connectedColor;
-        int blinkDuration;
-        int fadeDuration;
+        byte i2cDIR_Stored;
+        byte primaryColor;
+        byte secondaryColor;
+        byte disconnectedColor;
+        byte connectedColor;
+        byte blinkDuration;
+        byte fadeDuration;
         bool CH_On_0;
         bool CH_On_1;
         bool CH_On_2;
         bool CH_On_3;
-        int CH_Thr_0;
-        int CH_Thr_1;
-        int CH_Thr_2;
-        int CH_Thr_3;
+        byte CH_Thr_0;
+        byte CH_Thr_1;
+        byte CH_Thr_2;
+        byte CH_Thr_3;
     };
 
     storedConfiguration CONF;
